Fix for foliage animation appearing on items. Probably fixes some other inconsistencies.

+++ src/main/java/grondag/canvas/texture/MaterialIndexProvider.java
@@ -110,7 +110,7 @@ public abstract class MaterialIndexProvider {
 
 		@Override
 		public MaterialIndexer getIndexer(CanvasRenderMaterial mat) {
-			final long key = mat.vertexShaderIndex() | (mat.fragmentShaderIndex() << 16) | ((mat.shaderFlags()) << 32) | (((long) mat.condition().index()) << 48);
+			final long key = mat.vertexShaderIndex() | (mat.fragmentShaderIndex() << 16) | (((long) mat.shaderFlags()) << 32) | (((long) mat.condition().index()) << 48);
 
 			synchronized (sync) {
 				return materialMap.computeIfAbsent(key, k -> {
